+++ test/unit/fuzz-tools.test.js
@@ -353,6 +353,7 @@ import * as niRadsV373 from '../../lib/ni-rads-v373.js';
 import * as pauwelsFemoralNeckV374 from '../../lib/pauwels-femoral-neck-v374.js';
 import * as pipkinFemoralHeadV375 from '../../lib/pipkin-femoral-head-v375.js';
 import * as denisSacralV376 from '../../lib/denis-sacral-v376.js';
+import * as gartlandSupracondylarV377 from '../../lib/gartland-supracondylar-v377.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -651,6 +652,7 @@ const MODULES = {
   'pauwels-femoral-neck-v374.js': pauwelsFemoralNeckV374,
   'pipkin-femoral-head-v375.js': pipkinFemoralHeadV375,
   'denis-sacral-v376.js': denisSacralV376,
+  'gartland-supracondylar-v377.js': gartlandSupracondylarV377,
   'tb-testing.js': tbTesting,
 };